% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_pspm.R
\name{bootstrap_pspm}
\alias{bootstrap_pspm}
\title{PSPM Parametric Bootstrap}
\usage{
bootstrap_pspm(
  model = NULL,
  PSPMLearn = NULL,
  cl = NULL,
  n_boot_iter = 100,
  burnin = 100,
  ci_level = 0.95,
  return_sims = FALSE,
  cache_samples = T,
  beta = NULL
)
}
\arguments{
\item{model}{A model as returned by fit_pspm_model(..., return_pspm = TRUE). 
Must include a PSPMLearn object called "learn_obj". Can be NULL if PSPMLearn object is provided directly.}

\item{PSPMLearn}{A PSPMLearn object.  Can be NULL if model is provided instead.}

\item{cl}{CPU cluster object (created through \code{parallel::makeCluster}) 
or integer number of core. If the latter, a cluster is setup internally and
closed upon completion.}

\item{n_boot_iter}{Number of bootstrap iterations.}

\item{burnin}{Number of burn-in period for each sampling.}

\item{ci_level}{Confidence interval level.}

\item{return_sims}{Return all sampled partitionings?}

\item{cache_samples}{Cache the sampled partitionings in the PSPM objects inside the PSPMLearn object.}

\item{beta}{Defaults to NULL. Can be a vector of beta parameters to reset 
beta parameters in the model and associated PSPMLearn object. If fitted, the latter
beta parameters in the PSPMLearn objects are the fitted parameters. 
The function resets beta parameters to its old values when finished. 
To avoid inconsistencies inside PSPMLearn objects, 
samples cannot be cached if beta parameter is specified.}
}
\value{
A list with estimated parameters, confidence intervals, and 
possibly sampled partitionings.
}
\description{
Returns confidence intervals and full set of 
bootstrapped estimates from a parametric bootstrap of a fitted PSPM model.
}
