% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspm_R6.R
\name{clusterExport_fast}
\alias{clusterExport_fast}
\title{Fast Export of R-Objects to CPU Node Cluster}
\usage{
clusterExport_fast(cl, varlist, envir = .GlobalEnv)
}
\arguments{
\item{cl}{cluster object, created by \code{parallel::makeCluster}}

\item{varlist}{Character vector of object names}

\item{envir}{Environment from which to export objects. Defaults to \code{.GlobalEnv}.}
}
\value{
\code{TRUE}
}
\description{
Runs significantly faster than \code{parallel:clusterExport}, 
as it uses a temporary file that is written to disc and then read in parallel. 
This comes at the risk that the temporary file may survive if the function breaks
and that some types of objects, in particular when they include pointers, 
are broken upon arrival on the cluster.
}
