% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspm2texreg.R
\name{extract.pspmfit}
\alias{extract.pspmfit}
\title{Extension of texreg packages}
\usage{
extract.pspmfit(model, include.nobs = TRUE, include.loglik = TRUE, ...)
}
\arguments{
\item{model}{PSPM Model}

\item{include.nobs}{Include number of observations}

\item{include.loglik}{Include logliklihood}

\item{...}{Other paramters}
}
\value{
Object created by \code{texreg::createTexreg}
}
\description{
Extension of texreg packages
}
