% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pspm.R
\name{fit_pspm_model}
\alias{fit_pspm_model}
\title{Fit a PSPM Model on a list of graphs}
\usage{
fit_pspm_model(
  formula,
  g_ls,
  network_stat = list(),
  model_type = "composite_log_likelihood",
  sigma2 = 10,
  na.rm = TRUE,
  force_contiguous = TRUE,
  split_connect_comp = FALSE,
  min_component_size = 2,
  vertex_coords = NULL,
  return_pspm = FALSE,
  return_g_ls = FALSE
)
}
\arguments{
\item{formula}{Regression formula\code{y ~ x1 + x2}. y must refer to a vertex-attribut that 
encodes partition memberships. x1, x2, etc. must refer to edge-level attributes that 
encode edge-wise predictors.}

\item{g_ls}{List of igraph objects}

\item{network_stat}{(Empty) list of network statistics. Not used, for future extensions and backwards compatibility.}

\item{model_type}{Type of likelihood used. Only \code{"composite_log_likelihood"} 
is implemented at the moment but future extensions are possible.}

\item{sigma2}{Penalization parameter. Defaults to 10.}

\item{na.rm}{Remove missings from the data (both vertices and edges).}

\item{force_contiguous}{Enforce partitions' contiguity by resetting partition value such that 
non-contiguous partitions are split.}

\item{split_connect_comp}{Split graph into connected components. This maybe necessary for graphs with unconnected components.}

\item{min_component_size}{Minimum number of vertices per connected components 
(can be used to delete disconnected "islands" with only one vertex which 
by definition form their own partition if partitions are enforces to be contiguous)}

\item{vertex_coords}{Input spatial vertex coordinates to keep alongside the model. 
As matrix or SpatialPoints* object. If \code{na.rm = T}, 
coordinates will be deleted where vertex information is missing.}

\item{return_pspm}{Return PSPMLearn object alongside MaxLik model. Necessary to bootstrap standard errors.}

\item{return_g_ls}{Return (reduced) graph objects alongside MaxLik model. Useful for plotting purposes.}
}
\value{
Returns a MaxLik model object.
}
\description{
Fit a PSPM Model on a list of graphs
}
