% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_grid_data}
\alias{generate_grid_data}
\title{Generates square lattice with sampled partitioning}
\usage{
generate_grid_data(
  N_sqrd,
  beta0,
  beta = NULL,
  dep_structure = c("von_neumann", "moore", "hexagonal"),
  burnin = 10,
  temperatures = NULL,
  swap_iter = NULL
)
}
\arguments{
\item{N_sqrd}{Integer; side length of the lattice.}

\item{beta0}{Baseline repulsion (positive) or attraction (negative) among nodes.}

\item{beta}{Effect of predictors. 
Each predictor separates the lattice into a left and right part. Parts do not align.}

\item{dep_structure}{Dependency structure. 
One of \code{c("von_neumann", "moore", "hexagonal")}}

\item{burnin}{Length of burn-in period during sampling.}

\item{temperatures}{Not used. For potential parallel tempering.}

\item{swap_iter}{Not used. For potential parallel tempering with swapping.}
}
\value{
A PSPM object based on sqare lattice of side-length \code{N_sqrd}.
}
\description{
Generates square lattice with sampled partitioning
}
