% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_square_grid_lattice}
\alias{generate_square_grid_lattice}
\title{Returns adjacency matrix for square lattice of side-length \code{N_sqrd}}
\usage{
generate_square_grid_lattice(
  N_sqrd,
  dep_structure = c("von_neumann", "moore", "hexagonal")
)
}
\arguments{
\item{N_sqrd}{Side-length of grid}

\item{dep_structure}{Dependency structure. One of \code{c("von_neumann", "moore", "hexagonal")}.}
}
\description{
Returns adjacency matrix for square lattice of side-length \code{N_sqrd}
}
