% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspm_R6.R
\name{get_basic_ci}
\alias{get_basic_ci}
\title{Get basic CI from bootstrapped estimates}
\usage{
get_basic_ci(bmat, bpoint, alpha)
}
\arguments{
\item{bmat}{Bootstrapped coefficient matrix}

\item{bpoint}{Point estimate}

\item{alpha}{Confidence level}
}
\description{
Get basic CI from bootstrapped estimates
}
\keyword{internal}
