% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspm_R6.R
\name{get_percentile_ci}
\alias{get_percentile_ci}
\title{Get percentile-based CI from bootstrapped estimates}
\usage{
get_percentile_ci(bmat, alpha)
}
\arguments{
\item{bmat}{Bootstrapped coefficient matrix}

\item{alpha}{Confidence level}

\item{bpoint}{Point estimate}
}
\description{
Get percentile-based CI from bootstrapped estimates
}
\keyword{internal}
