% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{igraph2PSPM}
\alias{igraph2PSPM}
\title{Cast igraph as PSPM object}
\usage{
igraph2PSPM(
  g,
  outcome_name,
  edge_pred_names,
  network_stat = list(),
  vertex_pred_names = NULL,
  vertex_coords = NULL,
  verbose = FALSE,
  force_contiguous = TRUE
)
}
\arguments{
\item{g}{An igraph}

\item{outcome_name}{Vertex-level variable names as strings that encodes partition IDs, the main outcome.}

\item{edge_pred_names}{Edge-level variable names as strings that encode the predictors}

\item{network_stat}{List of supra-edge network statistics. 
Currently not used. For future backward compatibility.}

\item{vertex_pred_names}{List of vertex-level predictors. 
Currently not used. For future backward compatibility.}

\item{vertex_coords}{Vector of x and y coordinates of vertices. 
If provided, saves coordinates of vertices with PSPM object as \code{sp::SpatialPoints}.}

\item{verbose}{Verbosity}

\item{force_contiguous}{Force contiguous partitioning.}
}
\value{
Returns a PSPM object.
}
\description{
Cast igraph as PSPM object
}
