% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pspm.R
\name{model_graphframe}
\alias{model_graphframe}
\title{Reduce graph to data necessary to estimate PSPM}
\usage{
model_graphframe(
  g,
  formula,
  network_stat = list(),
  force_contiguous = T,
  na.rm = F
)
}
\arguments{
\item{g}{igraph object}

\item{formula}{Regression formula\code{y ~ x1 + x2}. y must refer to a vertex-attribute that 
encodes partition memberships. x1, x2, etc. must refer to edge-level attributes that 
encode edge-wise predictors.}

\item{network_stat}{(Empty) list of network statistics. Not used, for future extensions and backwards compatibility. 
Type of likelihood used. Only \code{"composite_log_likelihood"}}

\item{force_contiguous}{Whether to force partitions encoded by y to be contiguous.}

\item{na.rm}{Flag to remove vertices and edges with missing data from the graph.}
}
\value{
A clean graph to estimate a PSPM
}
\description{
Reduce graph to data necessary to estimate PSPM
}
