% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspm_R6.R
\name{PSPM}
\alias{PSPM}
\title{Probabilistic Spatial Partition Model Data Container}
\value{
Parameter vector, a numeric vector with a constant and 
one entry per edge-level predictor, in this order
}
\description{
Data container for a spatial network to be analysed with a Probabilistic Spatial Partition Model. 
Objects of this class are the main input to create a PSPMLearn object, 
which allows to fit the Probabilistic Partition Model to derive parameter estimates.
}
\details{
Write about details here.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{A}}{NxN adjacency matrix of baseline lattice}

\item{\code{g}}{A python object of class SpatialLattice}

\item{\code{pm}}{A python object of class PartitionModel}

\item{\code{N}}{Number of nodes in spatial lattice}

\item{\code{L}}{Number of edge-wise covariates}

\item{\code{Y}}{Vector of partition labels}

\item{\code{Z}}{List of length L of NxN edge-wise covariates}

\item{\code{Z.labs}}{Vector of edge-wise covariate labels}

\item{\code{P}}{Number of vertex predictors. Currently 1, not used. Only here for future extensions.}

\item{\code{X}}{NxP Matrix of vertex predictors, currently contains only one column full of 1s. 
Currently not used. Only here for future extensions.}

\item{\code{points}}{A matrix of xy coordinates or a SpatialPoints* object of nrow/length = N; 
corresponds to nodes.}

\item{\code{beta}}{Vector of beta parameters. Defaults to 0s}

\item{\code{net_stats}}{List of network statistics.}

\item{\code{verbose}}{Level of verbosity, binary flag.}

\item{\code{force_contiguous}}{Flag on whether to enforce contiguous partitions. Input by user.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PSPM-new}{\code{PSPM$new()}}
\item \href{#method-PSPM-set_beta}{\code{PSPM$set_beta()}}
\item \href{#method-PSPM-get_beta}{\code{PSPM$get_beta()}}
\item \href{#method-PSPM-plot_partitioning}{\code{PSPM$plot_partitioning()}}
\item \href{#method-PSPM-sample}{\code{PSPM$sample()}}
\item \href{#method-PSPM-get_composite_log_likelihood}{\code{PSPM$get_composite_log_likelihood()}}
\item \href{#method-PSPM-clone}{\code{PSPM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PSPM-new"></a>}}
\if{latex}{\out{\hypertarget{method-PSPM-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PSPM object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PSPM$new(
  Y,
  Z,
  X = NULL,
  A,
  net_stats = c("LatticeNeighborsNS"),
  beta = NULL,
  points = NULL,
  force_contiguous = TRUE,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Y}}{Integer vector of outcome partitioning.}

\item{\code{Z}}{List of weighted adjacency matrices, each encoding one edge-level predictor. 
Can be named, names are then recycled as variable names.}

\item{\code{X}}{Matrix of Vertex attributes. Not in use at the moment, 
for future extensions and backwards compatibility.}

\item{\code{A}}{Adjacency matrix}

\item{\code{net_stats}}{List of network statistics. 
Defaults to \code{c("LatticeNeighborsNS")} and included for 
for future extensions and backwards compatibility.}

\item{\code{beta}}{Preset beta parameters, if any.}

\item{\code{points}}{SpatialPoints object that give geolocation of vertices.}

\item{\code{force_contiguous}}{Flag to enforce contiguity of the graph 
and throw an error if that is not given.}

\item{\code{verbose}}{Flag for verbosity.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PSPM-set_beta"></a>}}
\if{latex}{\out{\hypertarget{method-PSPM-set_beta}{}}}
\subsection{Method \code{set_beta()}}{
Set beta parameters on PSPM object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PSPM$set_beta(beta)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{beta}}{Numeric vector with a constant and 
one entry per edge-level predictor, in this order}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PSPM-get_beta"></a>}}
\if{latex}{\out{\hypertarget{method-PSPM-get_beta}{}}}
\subsection{Method \code{get_beta()}}{
Get beta parameters on PSPM object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PSPM$get_beta()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PSPM-plot_partitioning"></a>}}
\if{latex}{\out{\hypertarget{method-PSPM-plot_partitioning}{}}}
\subsection{Method \code{plot_partitioning()}}{
Plot partitioning of PSPM object as a labelled graph.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PSPM$plot_partitioning(edge_predictor = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{edge_predictor}}{ID or name of edge predictor to include
to plot different edge types. Only binary edge predictors supported at the moment.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PSPM-sample"></a>}}
\if{latex}{\out{\hypertarget{method-PSPM-sample}{}}}
\subsection{Method \code{sample()}}{
Sample partitioning
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PSPM$sample(burnin = 0, return_full = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{burnin}}{Length of burn-in period}

\item{\code{return_full}}{Flag to return all partitionings, one for each burn-in iteration.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
If \code{return_full = TRUE}, a matrix of partition IDs. 
If \code{return_full = FALSE}, the partitioning of the PSPM object
is updated internally.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PSPM-get_composite_log_likelihood"></a>}}
\if{latex}{\out{\hypertarget{method-PSPM-get_composite_log_likelihood}{}}}
\subsection{Method \code{get_composite_log_likelihood()}}{
Compute the log composite likelihood
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PSPM$get_composite_log_likelihood(beta)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{beta}}{Beta parameters for which to query log composite likelihood}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns log composite likelihood
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PSPM-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PSPM-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PSPM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
