% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{PSPM2igraph}
\alias{PSPM2igraph}
\title{Cast PSPM as igraph object}
\usage{
PSPM2igraph(pspm)
}
\arguments{
\item{pspm}{A PSPM object to be cast as an igraph.}
}
\value{
Returns an igraph object that encodes all the information stored in the 
PSPM object as vertex and edge-level attributes.
}
\description{
Cast PSPM as igraph object
}
