% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspm2texreg.R
\name{pspm2table}
\alias{pspm2table}
\title{Transfer fitted PSPM Model output to tex table}
\usage{
pspm2table(
  model.ls,
  type = "latex",
  add.stats = c("N_groups", "N_edges", "N_instances", "MutInfo"),
  custom.gof.rows = NULL,
  stars = c(0.01, 0.05, 0.1),
  bootci = F,
  boottype = "percentile",
  ...
)
}
\arguments{
\item{model.ls}{List of models returned from \code{fit_pspm_model} or \code{PSPMLearn$fit_composite_log_likelihood}}

\item{type}{One of latex, text or html}

\item{add.stats}{Statistics to add.}

\item{custom.gof.rows}{Add custom rows.}

\item{stars}{Stars to add (only if native standard errors are used)}

\item{bootci}{List of bootstrapped CIs as returns from \code{bootstrap_pspm_model} or  or \code{PSPMLearn$bootstrap_composite_log_likelihood}}

\item{boottype}{Type of bootstrapped standard error (basic or percentile)}

\item{...}{Additional inputs taken by texreg (latex) / screenreg (text) / htmlreg (html), 
depending on \code{type}}
}
\description{
Builds directly on Philip Leifeld's texreg package.
}
