% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspm_R6.R
\name{pspm_set_seed}
\alias{pspm_set_seed}
\title{Set a random seed in R and Python}
\usage{
pspm_set_seed(seed)
}
\arguments{
\item{seed}{a single value, interpreted as an integer}
}
\description{
Set a random seed in R and Python
}
