% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pspm.R
\name{recode_contig_partitioning}
\alias{recode_contig_partitioning}
\title{Recode non-contiguous partitioning to be contiguous}
\usage{
recode_contig_partitioning(graph, part_attr)
}
\arguments{
\item{graph}{igraph object}

\item{part_attr}{Vertex attribute that encodes partition membership.}
}
\value{
Numeric vector that encodes vertices partition IDs for contiguous partitions.
}
\description{
Recodes partition memberships such that every partition is contiguous on the graph.
}
