% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspm_R6.R
\name{reset_pspm}
\alias{reset_pspm}
\title{Reset PSPM object}
\usage{
reset_pspm(x)
}
\arguments{
\item{x}{A PSPM object to be reset.}
}
\value{
The reset PSPM object.
}
\description{
Reset PSPM object to restore pointers and internal Python objects. This function
has to be called after reloading a PSPM object from disc or exporting it to a CPU cluster.
}
