% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspm_R6.R
\name{reset_pspmlearn}
\alias{reset_pspmlearn}
\title{Reset PSPMLearn object}
\usage{
reset_pspmlearn(x)
}
\arguments{
\item{x}{A PSPMLearn object to be reset.}
}
\value{
The reset PSPMLearn object.
}
\description{
Reset PSPMLearn object to restore pointers and internal Python objects. This function
has to be called after reloading a PSPMLearn object from disc or exporting it to a CPU cluster.
}
