% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pspm.R
\name{fit_mutualinfo}
\alias{fit_mutualinfo}
\title{Computes the normalized mutual information}
\usage{
fit_mutualinfo(X, Y)
}
\value{
A number between 0 and 1.
}
\description{
Computes the normalized mutual information
}
\details{
Inputs are defined as
     \code{X}: Predicted partitioning
     
     \code{Y}: Observed partitioning
}
