% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pspm.R
\name{get_fit_statistic}
\alias{get_fit_statistic}
\title{Compute any fit statistic for predicted partitionings stored in a learn object
(produced during parametric bootstrap).}
\usage{
get_fit_statistic(learn_obj, fun, return_boot = FALSE)
}
\value{
If \code{return_boot = TRUE}, the full distribution across all bootstrap iterations, 
otherwise their mean.
}
\description{
Compute any fit statistic for predicted partitionings stored in a learn object
(produced during parametric bootstrap).
}
\details{
Inputs are defined as
     \code{learn_obj}: As returned from \code{learn_obj$bootstrap_composite_log_likelihood} or 
     \code{learn_obj$par_bootstrap_composite_log_likelihood}
     
     \code{fun}: Function that computes fit statistic for partitionings, e.g. \code{fit_mutualinfo}
}
