% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{is_separated_by_col}
\alias{is_separated_by_col}
\title{Checks whether units i and k are separated by sep_col in a 
grid with side length N_sqrd}
\usage{
is_separated_by_col(i, j, N_sqrd, sep_col)
}
\description{
Checks whether units i and k are separated by sep_col in a 
grid with side length N_sqrd
}
\keyword{internal}
