% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{is_von_neumann_neighbor}
\alias{is_von_neumann_neighbor}
\title{Checks whether nodes i and j are von Neumann neighbors on a square grid
with side length N_sqrd.}
\usage{
is_von_neumann_neighbor(i, j, N_sqrd)
}
\description{
Checks whether nodes i and j are von Neumann neighbors on a square grid
with side length N_sqrd.
}
\keyword{internal}
